installer image: fixed password support disabled by default
Instead of hard-coding the password in all builds and letting the user
who calls image-installer decided whether he wants to use it, the
password is not set anymore and now has to be chosen via
local.conf. To ensure that the user really means it, the password then
has to be given explicitly to image-installer via the FIXED_PASSWORD
env variable.

The intention is to make it harder to accidentally install in this
insecure mode.

For testing purposes, development builds still have "refkit" set
as password.

+++ meta-refkit/recipes-image/images/initramfs-framework-refkit-luks.bb
@@ -105,7 +105,8 @@ refkit_luks () {
                             fatal "Error locking NVRAM area with index ${REFKIT_DISK_ENCRYPTION_NVRAM_INDEX}"
                         fi
                     fi
-                    if [ ! -s "$keyfile" ]; then
+                    if [ ! -s "$keyfile" ] &&
+                       [ "${REFKIT_DISK_ENCRYPTION_PASSWORD}" ]; then
                         printf "%s" "${REFKIT_DISK_ENCRYPTION_PASSWORD}" >"$keyfile"
                         keyfile_offset=0
                     fi
